pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smadex-ltv-prediction"
version = "1.0.0"
description = "Production-ready LTV prediction system for the Smadex Datathon"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
keywords = ["ltv", "prediction", "smadex", "mobile-apps"]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "dask[dataframe]>=2023.0.0",
    "pyarrow>=12.0.0",
    "fastparquet>=2023.0.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "catboost>=1.2.0",
    "torch>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    "streamlit>=1.28.0",
    "pyyaml>=6.0",
    "joblib>=1.3.0",
    "tqdm>=4.65.0",
    "jupyter>=1.0.0",
    "ipython>=8.12.0",
    "black>=23.0.0",
    "pytest>=7.3.0",
    "rich>=13.0.0",
    "psutil>=5.9.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
